fix: T6-193: Adjust unique email constraint for external users.
When a user logins via SSO with an email that's already taken by
a user internally, this shouldn't cause a problem because we
can use the unique external id as a distinguishing element.
This patch adjusts the User model to add an index which stores
the email, and externalId, and ensures that a uniqueness constraint
is upheld.